improve waterfall state structure
